WIPrefactor(favoritesUseCases):Move analytics tracking for update event
