+++ shared/src/commonMain/kotlin/com/mbta/tid/mbta_app/usecases/FavoritesUsecases.kt
@@ -1,25 +1,35 @@
 package com.mbta.tid.mbta_app.usecases
 
+import com.mbta.tid.mbta_app.analytics.Analytics
 import com.mbta.tid.mbta_app.model.RouteStopDirection
 import com.mbta.tid.mbta_app.repositories.IFavoritesRepository
 import org.koin.core.component.KoinComponent
 
-class FavoritesUsecases(private val repository: IFavoritesRepository) : KoinComponent {
+class FavoritesUsecases(
+    private val repository: IFavoritesRepository,
+    private val analytics: Analytics,
+) : KoinComponent {
 
     suspend fun getRouteStopDirectionFavorites(): Set<RouteStopDirection> {
         val storedFavorites = repository.getFavorites()
         return storedFavorites.routeStopDirection ?: emptySet()
     }
 
-    // Boolean return value indicates saved state
-    suspend fun toggleRouteStopDirection(route: String, stop: String, direction: Int): Boolean {
-        val routeStopDirection = RouteStopDirection(route, stop, direction)
-        return toggleRouteStopDirection(routeStopDirection)
-    }
-
-    suspend fun updateRouteStopDirections(newValues: Map<RouteStopDirection, Boolean>) {
+    suspend fun updateRouteStopDirections(
+        newValues: Map<RouteStopDirection, Boolean>,
+        context: EditFavoritesContext,
+        defaultDirection: Int,
+    ) {
         val storedFavorites = repository.getFavorites()
         val currentFavorites = (storedFavorites.routeStopDirection ?: emptySet()).toMutableSet()
+
+        val changedFavorites =
+            newValues.filter {
+                (!it.value && currentFavorites.contains(it.key)) ||
+                    (it.value && !(currentFavorites.contains(it.key)))
+            }
+
+        analytics.favoritesUpdated(changedFavorites, context, defaultDirection)
 
         newValues.forEach { (routeStopDirection, isFavorite) ->
             if (isFavorite) {
@@ -30,32 +40,9 @@ class FavoritesUsecases(private val repository: IFavoritesRepository) : KoinComp
         }
         repository.setFavorites(storedFavorites.copy(routeStopDirection = currentFavorites))
     }
+}
 
-    // Boolean return value indicates saved state
-    suspend fun toggleRouteStopDirection(routeStopDirection: RouteStopDirection): Boolean {
-        val storedFavorites = repository.getFavorites()
-        val currentFavorites = (storedFavorites.routeStopDirection ?: emptySet()).toMutableSet()
-        val containsFavorite = currentFavorites.contains(routeStopDirection)
-        if (containsFavorite) {
-            currentFavorites.remove(routeStopDirection)
-        } else {
-            currentFavorites.add(routeStopDirection)
-        }
-        repository.setFavorites(storedFavorites.copy(routeStopDirection = currentFavorites))
-        return !containsFavorite
-    }
-
-    suspend fun addRouteStopDirections(routeStopDirections: Set<RouteStopDirection>) {
-        val storedFavorites = repository.getFavorites()
-        val currentFavorites = (storedFavorites.routeStopDirection ?: emptySet()).toMutableSet()
-        currentFavorites.addAll(routeStopDirections)
-        repository.setFavorites(storedFavorites.copy(routeStopDirection = currentFavorites))
-    }
-
-    suspend fun removeRouteStopDirections(routeStopDirections: Set<RouteStopDirection>) {
-        val storedFavorites = repository.getFavorites()
-        val currentFavorites = (storedFavorites.routeStopDirection ?: emptySet()).toMutableSet()
-        currentFavorites.removeAll(routeStopDirections)
-        repository.setFavorites(storedFavorites.copy(routeStopDirection = currentFavorites))
-    }
+enum class EditFavoritesContext {
+    Favorites,
+    StopDetails,
 }
